+++ R/ggstats-package.R
@@ -11,7 +11,7 @@
 ## usethis namespace: end
 NULL
 
-utils::globalVariables(c("prop"))
+utils::globalVariables(c("prop", "count"))
 
 # \lifecycle{experimental}
 # \lifecycle{maturing}
@@ -31,4 +31,16 @@ utils::globalVariables(c("prop"))
   } else {
     b
   }
+}
+
+# helper for changing default aes
+# @param Stat a ggproto stat
+# @param ... new default aes
+#' @importFrom utils modifyList
+.modify_default_aes <- function(Stat, ...) {
+  new_aes <- ggplot2::aes(...)
+  Stat$default_aes <- ggplot2::aes(
+    !!!utils::modifyList(StatProp$default_aes, new_aes)
+  )
+  Stat
 }